models/ethical_safety.py
import torch
import torch.nn as nn
from typing import Dict, Tuple, List

class EthicalSafety(nn.Module):
    def __init__(self, hidden_dim: int):
        super().__init__()
        self.hidden_dim = hidden_dim
        
        # Ethical constraint encoder
        self.constraint_encoder = nn.Sequential(
            nn.Linear(hidden_dim, hidden_dim),
            nn.ReLU(),
            nn.Linear(hidden_dim, hidden_dim)
        )
        
        # Safety verification layers
        self.safety_check = nn.Sequential(
            nn.Linear(hidden_dim, hidden_dim // 2),
            nn.ReLU(),
            nn.Linear(hidden_dim // 2, 1),
            nn.Sigmoid()
        )

        # Ethical decision scorer
        self.ethical_scorer = nn.Sequential(
            nn.Linear(hidden_dim * 2, hidden_dim),
            nn.ReLU(),
            nn.Linear(hidden_dim, 1),
            nn.Sigmoid()
        )

        # Define basic ethical constraints
        self.ethical_constraints = [
            "do_no_harm",
            "respect_autonomy",
            "protect_privacy",
            "ensure_fairness",
            "maintain_transparency"
        ]

    def check_safety(self, state: torch.Tensor) -> Tuple[torch.Tensor, Dict]:
        """Verify if the current state meets safety requirements"""
        safety_score = self.safety_check(state)
        is_safe = safety_score > 0.5
        
        return is_safe, {
            'safety_score': safety_score,
            'safety_threshold': 0.5
        }

    def evaluate_ethics(self, action: torch.Tensor, context: torch.Tensor) -> Tuple[torch.Tensor, Dict]:
        """Evaluate ethical implications of an action"""
        combined = torch.cat([action, context], dim=-1)
        ethics_score = self.ethical_scorer(combined)
        
        return ethics_score > 0.7, {
            'ethics_score': ethics_score,
            'ethics_threshold': 0.7
        }

    def forward(self, state: torch.Tensor, action: torch.Tensor, context: torch.Tensor) -> Dict:
        """
        Perform ethical and safety evaluation
        Returns dict with safety checks and ethical assessments
        """
        # Encode current state against ethical constraints
        encoded_state = self.constraint_encoder(state)
        
        # Perform safety checks
        is_safe, safety_metrics = self.check_safety(encoded_state)
        
        # Evaluate ethical implications
        is_ethical, ethics_metrics = self.evaluate_ethics(action, context)
        
        return {
            'is_safe': is_safe,
            'is_ethical': is_ethical,
            'safety_metrics': safety_metrics,
            'ethics_metrics': ethics_metrics,
            'constraints_satisfied': torch.all(is_safe & is_ethical)
        }

    def mitigate_risks(self, action: torch.Tensor, safety_metrics: Dict) -> torch.Tensor:
        """Apply safety constraints to modify risky actions"""
        is_safe = safety_metrics.get('is_safe', True)
        if isinstance(is_safe, bool):
            is_safe_tensor = torch.full((action.size(0),), is_safe, dtype=torch.bool, device=action.device)
        else:
            is_safe_tensor = is_safe.squeeze(-1)
        unsafe_mask = ~is_safe_tensor
        scaled_action = action.clone()
        safety_score = safety_metrics.get('safety_score', torch.ones_like(action))
        scaled_action[unsafe_mask] *= safety_score[unsafe_mask]
        return scaled_action
